Create car-rental-system.cpp
This is a C++ program that simulates a car rental system and allows the user to rent a car from a car rental agency. It is an online software that enables the customer to view the available vehicles, register, and then book the car. The interface is user-friendly and allows for easy checking. It acts as a tool to save time, money, and labor during the rental process for corporations using object-oriented programming.​